refactor(catalog): extract postgres schema infrastructure
