+++ backend/app/services/component_catalog_service_postgres.py
@@ -11,24 +11,27 @@ from app.services.catalog.postgres_runtime import (
     _postgres_dsn,
     _split_sql_script,
 )
+from app.services.catalog.postgres_integrity import (
+    POSTGRES_INTEGRITY_GUARDS_VERSION,
+    POSTGRES_SEARCH_VERSION,
+    ensure_postgres_integrity_guards,
+    ensure_postgres_search_indexes,
+)
+from app.services.catalog.postgres_projections import (
+    POSTGRES_HEAD_PROJECTION_VERSION,
+    POSTGRES_REMOTE_HEAD_PROJECTION_VERSION,
+    ensure_component_heads_projection,
+    ensure_remote_component_heads_projection,
+)
+from app.services.catalog.postgres_schema import (
+    CATALOG_SCHEMA_EPOCH,
+    POSTGRES_SCHEMA_VERSION,
+    create_base_schema,
+)
 from app.services.catalog_schema_migrations import apply_catalog_migrations
 from app.services.component_catalog_domain import ComponentCatalogDomainService
 
 logger = logging.getLogger(__name__)
-
-
-# Written on every startup so an older Prism, which treats this row as a hard
-# precondition, can still open a database this build has touched. Schema changes
-# belong in app.services.catalog_schema_migrations, not here.
-POSTGRES_SCHEMA_VERSION = "catalog-postgres-v7"
-CATALOG_SCHEMA_EPOCH = "2"
-
-# Derived state. Each is rebuilt when its version changes, so these deliberately
-# stay outside the migration ladder, which records a migration as run once.
-POSTGRES_SEARCH_VERSION = "catalog-search-v3"
-POSTGRES_INTEGRITY_GUARDS_VERSION = "catalog-integrity-guards-v4"
-POSTGRES_HEAD_PROJECTION_VERSION = "catalog-component-heads-v5"
-POSTGRES_REMOTE_HEAD_PROJECTION_VERSION = "catalog-remote-heads-v4"
 
 
 class ComponentCatalogPostgresService(ComponentCatalogDomainService):
@@ -97,7 +100,7 @@ class ComponentCatalogPostgresService(ComponentCatalogDomainService):
                 # Prism reaching this database still finds the schema-version
                 # row it insists on, which is what keeps a rollback from
                 # needing a data restore.
-                self._create_schema(conn)
+                create_base_schema(conn)
                 conn.execute(
                     "INSERT INTO catalog_meta(key, value) VALUES (%s, %s) "
                     "ON CONFLICT(key) DO UPDATE SET value = EXCLUDED.value",
@@ -142,902 +145,17 @@ class ComponentCatalogPostgresService(ComponentCatalogDomainService):
 
     def _ensure_component_heads_projection(self, conn: CatalogPostgresConnection) -> None:
         """Install the current-head read model and its transactional refresh hooks."""
-        marker = conn.execute(
-            "SELECT value FROM catalog_meta WHERE key = %s",
-            ("postgres_head_projection_version",),
-        ).fetchone()
-        if marker and str(marker["value"]) == POSTGRES_HEAD_PROJECTION_VERSION:
-            return
-        conn.execute(
-            """
-            CREATE TABLE IF NOT EXISTS component_heads (
-                component_id TEXT PRIMARY KEY REFERENCES components(id) ON DELETE CASCADE,
-                revision_id TEXT NOT NULL,
-                slug TEXT NOT NULL,
-                source TEXT NOT NULL,
-                is_active INTEGER NOT NULL,
-                identity_kind TEXT NOT NULL,
-                stock_known INTEGER NOT NULL DEFAULT 0,
-                stock_quantity DOUBLE PRECISION NOT NULL,
-                stock_uom TEXT NOT NULL,
-                inventory_status TEXT NOT NULL,
-                version INTEGER NOT NULL,
-                workflow_stage TEXT NOT NULL,
-                name TEXT NOT NULL,
-                value TEXT NOT NULL,
-                description TEXT NOT NULL,
-                datasheet_url TEXT NOT NULL,
-                manufacturer TEXT NOT NULL,
-                mpn TEXT NOT NULL,
-                category TEXT NOT NULL,
-                package_name TEXT NOT NULL,
-                vendor TEXT NOT NULL,
-                vendor_part_number TEXT NOT NULL,
-                mass_g TEXT NOT NULL,
-                rqjc_c_w TEXT NOT NULL,
-                rqjc_top_c_w TEXT NOT NULL,
-                temp_max_c TEXT NOT NULL,
-                temp_min_c TEXT NOT NULL,
-                power_dissipation_w TEXT NOT NULL,
-                rate TEXT NOT NULL,
-                sap_code TEXT NOT NULL,
-                summary TEXT NOT NULL,
-                extra_fields TEXT NOT NULL,
-                search_document TEXT NOT NULL,
-                created_by TEXT NOT NULL,
-                revision_created_at TEXT NOT NULL,
-                updated_at TEXT NOT NULL,
-                has_symbol INTEGER NOT NULL DEFAULT 0,
-                has_footprint INTEGER NOT NULL DEFAULT 0,
-                symbol_library TEXT NOT NULL DEFAULT '',
-                symbol_name TEXT NOT NULL DEFAULT '',
-                footprint_library TEXT NOT NULL DEFAULT '',
-                footprint_name TEXT NOT NULL DEFAULT '',
-                default_representation_id TEXT NOT NULL DEFAULT '',
-                representation_count INTEGER NOT NULL DEFAULT 0,
-                symbol_variant_count INTEGER NOT NULL DEFAULT 0,
-                footprint_variant_count INTEGER NOT NULL DEFAULT 0,
-                inventory_sources TEXT NOT NULL DEFAULT '[]'
-            )
-            """
-        )
-        conn.execute(
-            "ALTER TABLE component_heads ADD COLUMN IF NOT EXISTS inventory_sources "
-            "TEXT NOT NULL DEFAULT '[]'"
-        )
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_component_heads_active_updated "
-            "ON component_heads(is_active, updated_at DESC)"
-        )
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_component_heads_workflow "
-            "ON component_heads(workflow_stage, updated_at DESC)"
-        )
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_component_heads_category "
-            "ON component_heads(category, updated_at DESC)"
-        )
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_component_heads_search_lower "
-            "ON component_heads(lower(search_document))"
-        )
-        conn.execute(
-            """
-            CREATE OR REPLACE FUNCTION prism_refresh_component_head(target_component_id TEXT)
-            RETURNS void
-            LANGUAGE plpgsql
-            SET search_path = catalog, public
-            AS $$
-            BEGIN
-                DELETE FROM component_heads WHERE component_id = target_component_id;
-                INSERT INTO component_heads (
-                    component_id, revision_id, slug, source, is_active, identity_kind, stock_known,
-                    stock_quantity, stock_uom,
-                    inventory_status, version, workflow_stage, name, value, description, datasheet_url,
-                    manufacturer, mpn, category, package_name, vendor, vendor_part_number, mass_g,
-                    rqjc_c_w, rqjc_top_c_w, temp_max_c, temp_min_c, power_dissipation_w, rate,
-                    sap_code, summary, extra_fields, search_document, created_by, revision_created_at,
-                    updated_at, has_symbol, has_footprint, symbol_library, symbol_name,
-                    footprint_library, footprint_name, default_representation_id,
-                    representation_count, symbol_variant_count, footprint_variant_count,
-                    inventory_sources
-                )
-                SELECT
-                    component.id, revision.id, component.slug, component.source, component.is_active,
-                    component.identity_kind, CASE WHEN inventory.source IS NULL THEN 0 ELSE 1 END,
-                    COALESCE(inventory.quantity, 0), COALESCE(inventory.uom, ''),
-                    COALESCE(inventory.inventory_status, ''),
-                    revision.version, revision.release_status, revision.name, revision.value,
-                    revision.description, revision.datasheet_url, revision.manufacturer, revision.mpn,
-                    revision.category, revision.package_name, revision.vendor,
-                    revision.vendor_part_number, revision.mass_g, revision.rqjc_c_w,
-                    revision.rqjc_top_c_w, revision.temp_max_c, revision.temp_min_c,
-                    revision.power_dissipation_w, revision.rate, revision.sap_code, revision.summary,
-                    revision.extra_fields, revision.search_document, revision.created_by,
-                    revision.created_at, revision.updated_at,
-                    CASE WHEN symbol.id IS NULL THEN 0 ELSE 1 END,
-                    CASE WHEN footprint.id IS NULL THEN 0 ELSE 1 END,
-                    COALESCE(symbol.target_library, ''), COALESCE(symbol.target_name, ''),
-                    COALESCE(footprint.target_library, ''), COALESCE(footprint.target_name, ''),
-                    COALESCE(representation.id, ''), COALESCE(counts.representation_count, 0),
-                    COALESCE(counts.symbol_variant_count, 0), COALESCE(counts.footprint_variant_count, 0),
-                    COALESCE(inventory_all.sources_json, '[]')
-                FROM components component
-                JOIN component_revisions revision ON revision.id = component.current_revision_id
-                LEFT JOIN revision_representations representation
-                  ON representation.revision_id = revision.id AND representation.is_default = 1
-                LEFT JOIN assets symbol ON symbol.id = representation.symbol_asset_id
-                LEFT JOIN assets footprint ON footprint.id = representation.footprint_asset_id
-                LEFT JOIN LATERAL (
-                    SELECT COUNT(*)::integer AS representation_count,
-                           COUNT(DISTINCT symbol_asset_id)::integer AS symbol_variant_count,
-                           COUNT(DISTINCT footprint_asset_id)::integer AS footprint_variant_count
-                    FROM revision_representations WHERE revision_id = revision.id
-                ) counts ON true
-                LEFT JOIN LATERAL (
-                    SELECT source, SUM(quantity) AS quantity, MIN(uom) AS uom,
-                           MIN(inventory_status) AS inventory_status
-                    FROM inventory_levels WHERE component_id = component.id
-                    GROUP BY source
-                    ORDER BY CASE source WHEN 'inventree' THEN 1 WHEN 'csv' THEN 2 ELSE 99 END
-                    LIMIT 1
-                ) inventory ON true
-                LEFT JOIN LATERAL (
-                    SELECT COALESCE(
-                        json_agg(json_build_object(
-                            'source', agg.source,
-                            'quantity', agg.quantity,
-                            'uom', agg.uom,
-                            'inventory_status', agg.inventory_status,
-                            'fetch_status', agg.fetch_status,
-                            'fetched_at', agg.fetched_at
-                        ) ORDER BY CASE agg.source WHEN 'inventree' THEN 1 WHEN 'csv' THEN 2 ELSE 99 END,
-                                  agg.source)::text, '[]') AS sources_json
-                    FROM (
-                        SELECT source, SUM(quantity) AS quantity, MIN(uom) AS uom,
-                               MIN(inventory_status) AS inventory_status,
-                               MIN(fetch_status) AS fetch_status,
-                               MAX(fetched_at) AS fetched_at
-                        FROM inventory_levels WHERE component_id = component.id
-                        GROUP BY source
-                    ) agg
-                ) inventory_all ON true
-                WHERE component.id = target_component_id AND component.current_revision_id <> '';
-            END;
-            $$
-            """
-        )
-        conn.execute(
-            """
-            CREATE OR REPLACE FUNCTION prism_refresh_component_head_trigger()
-            RETURNS trigger
-            LANGUAGE plpgsql
-            SET search_path = catalog, public
-            AS $$
-            DECLARE
-                target_component_id TEXT;
-            BEGIN
-                IF TG_TABLE_NAME = 'components' THEN
-                    target_component_id := CASE WHEN TG_OP = 'DELETE' THEN OLD.id ELSE NEW.id END;
-                ELSE
-                    target_component_id := CASE WHEN TG_OP = 'DELETE' THEN OLD.component_id ELSE NEW.component_id END;
-                END IF;
-                PERFORM catalog.prism_refresh_component_head(target_component_id);
-                RETURN CASE WHEN TG_OP = 'DELETE' THEN OLD ELSE NEW END;
-            END;
-            $$
-            """
-        )
-        conn.execute(
-            """
-            CREATE OR REPLACE FUNCTION prism_refresh_component_head_asset_trigger()
-            RETURNS trigger
-            LANGUAGE plpgsql
-            SET search_path = catalog, public
-            AS $$
-            DECLARE
-                target_revision_id TEXT;
-                target_component_id TEXT;
-            BEGIN
-                IF TG_TABLE_NAME = 'inventory_levels' THEN
-                    target_component_id := CASE WHEN TG_OP = 'DELETE' THEN OLD.component_id ELSE NEW.component_id END;
-                ELSE
-                    target_revision_id := CASE WHEN TG_OP = 'DELETE' THEN OLD.revision_id ELSE NEW.revision_id END;
-                    SELECT component_id INTO target_component_id
-                    FROM component_revisions WHERE id = target_revision_id;
-                END IF;
-                IF target_component_id IS NOT NULL THEN
-                    PERFORM catalog.prism_refresh_component_head(target_component_id);
-                END IF;
-                RETURN CASE WHEN TG_OP = 'DELETE' THEN OLD ELSE NEW END;
-            END;
-            $$
-            """
-        )
-        for trigger_name, table, events, function_name in (
-            ("trg_component_heads_components", "components", "INSERT OR UPDATE", "prism_refresh_component_head_trigger"),
-            ("trg_component_heads_revisions", "component_revisions", "INSERT OR UPDATE", "prism_refresh_component_head_trigger"),
-            ("trg_component_heads_assets", "revision_assets", "INSERT OR UPDATE OR DELETE", "prism_refresh_component_head_asset_trigger"),
-            ("trg_component_heads_representations", "revision_representations", "INSERT OR UPDATE OR DELETE", "prism_refresh_component_head_asset_trigger"),
-            ("trg_component_heads_inventory", "inventory_levels", "INSERT OR UPDATE OR DELETE", "prism_refresh_component_head_asset_trigger"),
-        ):
-            conn.execute(f"DROP TRIGGER IF EXISTS {trigger_name} ON {table}")
-            conn.execute(
-                f"CREATE TRIGGER {trigger_name} AFTER {events} ON {table} "
-                f"FOR EACH ROW EXECUTE FUNCTION {function_name}()"
-            )
-        conn.execute("SELECT catalog.prism_refresh_component_head(id) FROM components")
-        conn.execute(
-            """
-            INSERT INTO catalog_meta (key, value) VALUES (%s, %s)
-            ON CONFLICT(key) DO UPDATE SET value = excluded.value
-            """,
-            ("postgres_head_projection_version", POSTGRES_HEAD_PROJECTION_VERSION),
-        )
+        ensure_component_heads_projection(conn)
 
     def _ensure_remote_component_heads_projection(self, conn: CatalogPostgresConnection) -> None:
         """Install the released-only read model used by the KiCad provider."""
-
-        marker = conn.execute(
-            "SELECT value FROM catalog_meta WHERE key = %s",
-            ("postgres_remote_head_projection_version",),
-        ).fetchone()
-        if marker and str(marker["value"]) == POSTGRES_REMOTE_HEAD_PROJECTION_VERSION:
-            return
-        conn.execute(
-            """
-            CREATE TABLE IF NOT EXISTS remote_component_heads (
-                component_id TEXT PRIMARY KEY REFERENCES components(id) ON DELETE CASCADE,
-                revision_id TEXT NOT NULL,
-                slug TEXT NOT NULL,
-                source TEXT NOT NULL,
-                is_active INTEGER NOT NULL,
-                identity_kind TEXT NOT NULL,
-                stock_known INTEGER NOT NULL DEFAULT 0,
-                stock_quantity DOUBLE PRECISION NOT NULL,
-                stock_uom TEXT NOT NULL,
-                inventory_status TEXT NOT NULL,
-                version INTEGER NOT NULL,
-                name TEXT NOT NULL,
-                description TEXT NOT NULL,
-                datasheet_url TEXT NOT NULL,
-                manufacturer TEXT NOT NULL,
-                mpn TEXT NOT NULL,
-                category TEXT NOT NULL,
-                package_name TEXT NOT NULL,
-                summary TEXT NOT NULL,
-                extra_fields TEXT NOT NULL,
-                search_document TEXT NOT NULL,
-                updated_at TEXT NOT NULL,
-                has_symbol INTEGER NOT NULL DEFAULT 0,
-                has_footprint INTEGER NOT NULL DEFAULT 0,
-                symbol_library TEXT NOT NULL DEFAULT '',
-                symbol_name TEXT NOT NULL DEFAULT '',
-                symbol_preview_id TEXT NOT NULL DEFAULT '',
-                footprint_preview_id TEXT NOT NULL DEFAULT '',
-                footprint_library TEXT NOT NULL DEFAULT '',
-                footprint_name TEXT NOT NULL DEFAULT '',
-                default_representation_id TEXT NOT NULL DEFAULT '',
-                representation_count INTEGER NOT NULL DEFAULT 0,
-                symbol_variant_count INTEGER NOT NULL DEFAULT 0,
-                footprint_variant_count INTEGER NOT NULL DEFAULT 0,
-                inventory_sources TEXT NOT NULL DEFAULT '[]'
-            )
-            """
-        )
-        conn.execute(
-            "ALTER TABLE remote_component_heads ADD COLUMN IF NOT EXISTS inventory_sources "
-            "TEXT NOT NULL DEFAULT '[]'"
-        )
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_remote_component_heads_updated "
-            "ON remote_component_heads(updated_at DESC)"
-        )
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_remote_component_heads_category "
-            "ON remote_component_heads(category, updated_at DESC)"
-        )
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_remote_component_heads_search_lower "
-            "ON remote_component_heads(lower(search_document))"
-        )
-        conn.execute(
-            """
-            CREATE OR REPLACE FUNCTION prism_refresh_remote_component_head(
-                target_component_id TEXT
-            )
-            RETURNS void
-            LANGUAGE plpgsql
-            SET search_path = catalog, public
-            AS $$
-            BEGIN
-                DELETE FROM remote_component_heads
-                WHERE component_id = target_component_id;
-                INSERT INTO remote_component_heads (
-                    component_id, revision_id, slug, source, is_active, identity_kind, stock_known,
-                    stock_quantity, stock_uom, inventory_status, version,
-                    name, description, datasheet_url, manufacturer, mpn,
-                    category, package_name, summary, extra_fields,
-                    search_document, updated_at, has_symbol, has_footprint,
-                    symbol_library, symbol_name, symbol_preview_id,
-                    footprint_preview_id, footprint_library, footprint_name,
-                    default_representation_id, representation_count,
-                    symbol_variant_count, footprint_variant_count,
-                    inventory_sources
-                )
-                SELECT
-                    component.id, revision.id, component.slug, component.source,
-                    component.is_active, component.identity_kind,
-                    CASE WHEN inventory.source IS NULL THEN 0 ELSE 1 END,
-                    COALESCE(inventory.quantity, 0), COALESCE(inventory.uom, ''),
-                    COALESCE(inventory.inventory_status, ''),
-                    revision.version, revision.name, revision.description,
-                    revision.datasheet_url, revision.manufacturer, revision.mpn,
-                    revision.category, revision.package_name, revision.summary,
-                    revision.extra_fields, revision.search_document,
-                    revision.updated_at,
-                    CASE WHEN symbol.id IS NULL THEN 0 ELSE 1 END,
-                    CASE WHEN footprint.id IS NULL THEN 0 ELSE 1 END,
-                    COALESCE(symbol.target_library, ''),
-                    COALESCE(symbol.target_name, ''),
-                    COALESCE(symbol_preview.preview_id, ''),
-                    COALESCE(footprint_preview.preview_id, ''),
-                    COALESCE(footprint.target_library, ''), COALESCE(footprint.target_name, ''),
-                    COALESCE(representation.id, ''), COALESCE(counts.representation_count, 0),
-                    COALESCE(counts.symbol_variant_count, 0), COALESCE(counts.footprint_variant_count, 0),
-                    COALESCE(inventory_all.sources_json, '[]')
-                FROM components component
-                JOIN component_revisions revision
-                  ON revision.id = component.released_revision_id
-                LEFT JOIN revision_representations representation
-                  ON representation.revision_id = revision.id AND representation.is_default = 1
-                LEFT JOIN assets symbol ON symbol.id = representation.symbol_asset_id
-                LEFT JOIN assets footprint ON footprint.id = representation.footprint_asset_id
-                LEFT JOIN LATERAL (
-                    SELECT COUNT(*)::integer AS representation_count,
-                           COUNT(DISTINCT symbol_asset_id)::integer AS symbol_variant_count,
-                           COUNT(DISTINCT footprint_asset_id)::integer AS footprint_variant_count
-                    FROM revision_representations WHERE revision_id = revision.id
-                ) counts ON true
-                LEFT JOIN LATERAL (
-                    SELECT source, SUM(quantity) AS quantity, MIN(uom) AS uom,
-                           MIN(inventory_status) AS inventory_status
-                    FROM inventory_levels WHERE component_id = component.id
-                    GROUP BY source
-                    ORDER BY CASE source WHEN 'inventree' THEN 1 WHEN 'csv' THEN 2 ELSE 99 END
-                    LIMIT 1
-                ) inventory ON true
-                LEFT JOIN LATERAL (
-                    SELECT COALESCE(
-                        json_agg(json_build_object(
-                            'source', agg.source,
-                            'quantity', agg.quantity,
-                            'uom', agg.uom,
-                            'inventory_status', agg.inventory_status,
-                            'fetch_status', agg.fetch_status,
-                            'fetched_at', agg.fetched_at
-                        ) ORDER BY CASE agg.source WHEN 'inventree' THEN 1 WHEN 'csv' THEN 2 ELSE 99 END,
-                                  agg.source)::text, '[]') AS sources_json
-                    FROM (
-                        SELECT source, SUM(quantity) AS quantity, MIN(uom) AS uom,
-                               MIN(inventory_status) AS inventory_status,
-                               MIN(fetch_status) AS fetch_status,
-                               MAX(fetched_at) AS fetched_at
-                        FROM inventory_levels WHERE component_id = component.id
-                        GROUP BY source
-                    ) agg
-                ) inventory_all ON true
-                LEFT JOIN LATERAL (
-                    SELECT preview.id AS preview_id
-                    FROM revision_preview_outputs link
-                    JOIN asset_preview_versions preview
-                      ON preview.id = link.preview_id
-                    WHERE link.revision_id = revision.id
-                      AND link.asset_id = representation.symbol_asset_id
-                      AND link.kind = 'symbol'
-                      AND preview.status = 'ready'
-                      AND preview.file_path <> ''
-                    ORDER BY preview.created_at DESC
-                    LIMIT 1
-                ) symbol_preview ON true
-                LEFT JOIN LATERAL (
-                    SELECT preview.id AS preview_id
-                    FROM revision_preview_outputs link
-                    JOIN asset_preview_versions preview
-                      ON preview.id = link.preview_id
-                    WHERE link.revision_id = revision.id
-                      AND link.asset_id = representation.footprint_asset_id
-                      AND link.kind = 'footprint'
-                      AND preview.status = 'ready'
-                      AND preview.file_path <> ''
-                    ORDER BY preview.created_at DESC
-                    LIMIT 1
-                ) footprint_preview ON true
-                WHERE component.id = target_component_id
-                  AND component.is_active = 1
-                  AND component.released_revision_id <> ''
-                  AND revision.release_status = 'released';
-                INSERT INTO catalog_meta(key, value)
-                VALUES (
-                    'remote_component_heads_version',
-                    EXTRACT(EPOCH FROM clock_timestamp())::text
-                )
-                ON CONFLICT(key) DO UPDATE SET value = EXCLUDED.value;
-            END;
-            $$
-            """
-        )
-        conn.execute(
-            """
-            CREATE OR REPLACE FUNCTION prism_refresh_remote_head_trigger()
-            RETURNS trigger
-            LANGUAGE plpgsql
-            SET search_path = catalog, public
-            AS $$
-            DECLARE
-                target_component_id TEXT;
-                target_revision_id TEXT;
-            BEGIN
-                IF TG_TABLE_NAME = 'components' THEN
-                    target_component_id :=
-                        CASE WHEN TG_OP = 'DELETE' THEN OLD.id ELSE NEW.id END;
-                ELSIF TG_TABLE_NAME = 'component_revisions' THEN
-                    target_component_id :=
-                        CASE WHEN TG_OP = 'DELETE'
-                             THEN OLD.component_id ELSE NEW.component_id END;
-                ELSIF TG_TABLE_NAME = 'inventory_levels' THEN
-                    target_component_id :=
-                        CASE WHEN TG_OP = 'DELETE'
-                             THEN OLD.component_id ELSE NEW.component_id END;
-                ELSE
-                    target_revision_id :=
-                        CASE WHEN TG_OP = 'DELETE'
-                             THEN OLD.revision_id ELSE NEW.revision_id END;
-                    SELECT component_id INTO target_component_id
-                    FROM component_revisions
-                    WHERE id = target_revision_id;
-                END IF;
-                IF target_component_id IS NOT NULL THEN
-                    PERFORM catalog.prism_refresh_remote_component_head(
-                        target_component_id
-                    );
-                END IF;
-                RETURN CASE WHEN TG_OP = 'DELETE' THEN OLD ELSE NEW END;
-            END;
-            $$
-            """
-        )
-        for trigger_name, table, events in (
-            (
-                "trg_remote_heads_components",
-                "components",
-                "INSERT OR UPDATE",
-            ),
-            (
-                "trg_remote_heads_revisions",
-                "component_revisions",
-                "INSERT OR UPDATE",
-            ),
-            (
-                "trg_remote_heads_assets",
-                "revision_assets",
-                "INSERT OR UPDATE OR DELETE",
-            ),
-            (
-                "trg_remote_heads_representations",
-                "revision_representations",
-                "INSERT OR UPDATE OR DELETE",
-            ),
-            (
-                "trg_remote_heads_inventory",
-                "inventory_levels",
-                "INSERT OR UPDATE OR DELETE",
-            ),
-            (
-                "trg_remote_heads_previews",
-                "revision_preview_outputs",
-                "INSERT OR UPDATE OR DELETE",
-            ),
-        ):
-            conn.execute(f"DROP TRIGGER IF EXISTS {trigger_name} ON {table}")
-            conn.execute(
-                f"CREATE TRIGGER {trigger_name} AFTER {events} ON {table} "
-                "FOR EACH ROW EXECUTE FUNCTION prism_refresh_remote_head_trigger()"
-            )
-        conn.execute(
-            "SELECT catalog.prism_refresh_remote_component_head(id) FROM components"
-        )
-        conn.execute(
-            """
-            INSERT INTO catalog_meta(key, value)
-            VALUES (%s, %s)
-            ON CONFLICT(key) DO UPDATE SET value = EXCLUDED.value
-            """,
-            (
-                "postgres_remote_head_projection_version",
-                POSTGRES_REMOTE_HEAD_PROJECTION_VERSION,
-            ),
-        )
+        ensure_remote_component_heads_projection(conn)
 
     def _ensure_postgres_search_indexes(self) -> None:
-        # Trigram search keeps the existing forgiving catalog query behavior while
-        # avoiding full scans at tens of thousands of components. Extension creation
-        # can be disallowed on managed databases, so degrade to ordinary indexes.
-        with self._connect() as conn:
-            conn.execute(
-                "SELECT pg_advisory_xact_lock(hashtext(%s))",
-                ("prism-component-catalog-search",),
-            ).fetchone()
-            marker = conn.execute(
-                "SELECT value FROM catalog_meta WHERE key = %s",
-                ("postgres_search_version",),
-            ).fetchone()
-            if marker and str(marker["value"]) == POSTGRES_SEARCH_VERSION:
-                conn.commit()
-                return
-            try:
-                conn.execute("CREATE EXTENSION IF NOT EXISTS pg_trgm")
-                conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_revisions_search_trgm "
-                    "ON component_revisions USING GIN (lower(search_document) gin_trgm_ops)"
-                )
-                conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_revisions_mpn_trgm "
-                    "ON component_revisions USING GIN (lower(mpn) gin_trgm_ops)"
-                )
-                conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_remote_heads_search_trgm "
-                    "ON remote_component_heads USING GIN "
-                    "(lower(search_document) gin_trgm_ops)"
-                )
-                conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_remote_heads_mpn_trgm "
-                    "ON remote_component_heads USING GIN (lower(mpn) gin_trgm_ops)"
-                )
-                # The asset link picker searches with leading wildcards, which no
-                # btree index can serve, so without these it sequentially scans
-                # every asset on each keystroke.
-                conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_assets_name_trgm "
-                    "ON assets USING GIN (lower(name) gin_trgm_ops)"
-                )
-                conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_assets_target_name_trgm "
-                    "ON assets USING GIN (lower(target_name) gin_trgm_ops)"
-                )
-                conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_assets_target_library_trgm "
-                    "ON assets USING GIN (lower(target_library) gin_trgm_ops)"
-                )
-                # A new index is invisible to the planner until the table has
-                # statistics that justify it: measured on ~17k assets, the search
-                # kept seq scanning until this ran, then dropped from 27ms to
-                # under 1ms. Waiting for autovacuum would leave every deploy slow
-                # for as long as that takes to come around.
-                conn.execute("ANALYZE assets")
-                conn.execute(
-                    """
-                    INSERT INTO catalog_meta (key, value) VALUES (%s, %s)
-                    ON CONFLICT (key) DO UPDATE SET value = EXCLUDED.value
-                    """,
-                    ("postgres_search_version", POSTGRES_SEARCH_VERSION),
-                )
-                conn.commit()
-            except Exception as exc:
-                logger.warning(
-                    "pg_trgm catalog search indexes unavailable; falling back to btree lower() indexes: %s",
-                    exc,
-                )
-                conn.rollback()
-                conn.execute(
-                    "SELECT pg_advisory_xact_lock(hashtext(%s))",
-                    ("prism-component-catalog-search",),
-                ).fetchone()
-                conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_revisions_search_lower "
-                    "ON component_revisions(lower(search_document))"
-                )
-                conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_revisions_mpn_lower "
-                    "ON component_revisions(lower(mpn))"
-                )
-                conn.execute(
-                    """
-                    INSERT INTO catalog_meta (key, value) VALUES (%s, %s)
-                    ON CONFLICT (key) DO UPDATE SET value = EXCLUDED.value
-                    """,
-                    ("postgres_search_version", POSTGRES_SEARCH_VERSION),
-                )
-                conn.commit()
+        ensure_postgres_search_indexes(self._postgres_runtime)
 
     def _ensure_postgres_integrity_guards(self) -> None:
-        with self._connect() as conn:
-            conn.execute(
-                "SELECT pg_advisory_xact_lock(hashtext(%s))",
-                ("prism-component-catalog-integrity-guards",),
-            ).fetchone()
-            marker = conn.execute(
-                "SELECT value FROM catalog_meta WHERE key = %s",
-                ("postgres_integrity_guards_version",),
-            ).fetchone()
-            if marker and str(marker["value"]) == POSTGRES_INTEGRITY_GUARDS_VERSION:
-                conn.commit()
-                return
-            conn.execute(
-                """
-                CREATE OR REPLACE FUNCTION prism_reject_catalog_evidence_mutation()
-                RETURNS trigger
-                LANGUAGE plpgsql
-                AS $$
-                BEGIN
-                    RAISE EXCEPTION 'immutable catalog evidence cannot be updated or deleted';
-                END;
-                $$
-                """
-            )
-            guarded_tables = {
-                "catalog_audit_events": "UPDATE OR DELETE",
-                "component_review_decisions": "UPDATE OR DELETE",
-                "component_release_records": "UPDATE OR DELETE",
-                "components": "DELETE",
-                "component_revisions": "DELETE",
-                "asset_previews": "UPDATE OR DELETE",
-                "asset_preview_versions": "UPDATE OR DELETE",
-            }
-            for table, operations in guarded_tables.items():
-                trigger_name = f"trg_{table}_immutable"
-                exists = conn.execute(
-                    """
-                    SELECT 1 AS present
-                    FROM pg_trigger
-                    WHERE tgname = %s AND tgrelid = to_regclass(%s) AND NOT tgisinternal
-                    """,
-                    (trigger_name, f"catalog.{table}"),
-                ).fetchone()
-                if not exists:
-                    conn.execute(
-                        f"CREATE TRIGGER {trigger_name} BEFORE {operations} ON {table} "
-                        "FOR EACH ROW EXECUTE FUNCTION prism_reject_catalog_evidence_mutation()"
-                    )
-            conn.execute(
-                """
-                CREATE OR REPLACE FUNCTION prism_guard_revision_preview_mutation()
-                RETURNS trigger
-                LANGUAGE plpgsql
-                AS $$
-                DECLARE
-                    guarded_revision_id TEXT;
-                    parent_manifest_hash TEXT;
-                BEGIN
-                    IF current_setting('prism.catalog_migration', true) = 'on' THEN
-                        RETURN CASE WHEN TG_OP = 'DELETE' THEN OLD ELSE NEW END;
-                    END IF;
-                    guarded_revision_id := CASE WHEN TG_OP = 'DELETE' THEN OLD.revision_id ELSE NEW.revision_id END;
-                    SELECT manifest_hash INTO parent_manifest_hash
-                    FROM component_revisions revision
-                    WHERE revision.id = guarded_revision_id;
-                    IF COALESCE(parent_manifest_hash, '') <> '' THEN
-                        RAISE EXCEPTION 'finalized revision preview evidence is immutable';
-                    END IF;
-                    RETURN CASE WHEN TG_OP = 'DELETE' THEN OLD ELSE NEW END;
-                END;
-                $$
-                """
-            )
-            conn.execute(
-                """
-                CREATE OR REPLACE FUNCTION prism_guard_finalized_revision_update()
-                RETURNS trigger
-                LANGUAGE plpgsql
-                AS $$
-                BEGIN
-                    IF current_setting('prism.catalog_migration', true) = 'on' THEN
-                        RETURN NEW;
-                    END IF;
-                    IF COALESCE(OLD.manifest_hash, '') <> ''
-                       AND (to_jsonb(NEW) - ARRAY['release_status', 'updated_at'])
-                           IS DISTINCT FROM
-                           (to_jsonb(OLD) - ARRAY['release_status', 'updated_at']) THEN
-                        RAISE EXCEPTION 'finalized component revision evidence is immutable';
-                    END IF;
-                    RETURN NEW;
-                END;
-                $$
-                """
-            )
-            conn.execute(
-                """
-                CREATE OR REPLACE FUNCTION prism_guard_asset_identity_update()
-                RETURNS trigger
-                LANGUAGE plpgsql
-                AS $$
-                BEGIN
-                    IF current_setting('prism.catalog_migration', true) = 'on' THEN
-                        RETURN NEW;
-                    END IF;
-                    IF (to_jsonb(NEW) - ARRAY['name', 'canonical_path', 'size_bytes', 'content_type', 'updated_at'])
-                       IS DISTINCT FROM
-                       (to_jsonb(OLD) - ARRAY['name', 'canonical_path', 'size_bytes', 'content_type', 'updated_at']) THEN
-                        RAISE EXCEPTION 'immutable asset identity or content hash cannot be changed';
-                    END IF;
-                    RETURN NEW;
-                END;
-                $$
-                """
-            )
-            conn.execute(
-                """
-                CREATE OR REPLACE FUNCTION prism_validate_revision_asset_type()
-                RETURNS trigger
-                LANGUAGE plpgsql
-                AS $$
-                DECLARE actual_type TEXT;
-                BEGIN
-                    SELECT asset_type INTO actual_type FROM assets WHERE id = NEW.asset_id;
-                    IF actual_type IS DISTINCT FROM NEW.asset_type THEN
-                        RAISE EXCEPTION 'revision asset type does not match asset';
-                    END IF;
-                    RETURN NEW;
-                END;
-                $$
-                """
-            )
-            conn.execute(
-                """
-                CREATE OR REPLACE FUNCTION prism_validate_representation_assets()
-                RETURNS trigger
-                LANGUAGE plpgsql
-                AS $$
-                BEGIN
-                    IF NEW.symbol_asset_id IS NOT NULL AND NOT EXISTS (
-                        SELECT 1 FROM revision_assets link JOIN assets asset ON asset.id = link.asset_id
-                        WHERE link.revision_id = NEW.revision_id AND link.asset_id = NEW.symbol_asset_id
-                          AND link.asset_type = 'symbol' AND asset.asset_type = 'symbol'
-                    ) THEN
-                        RAISE EXCEPTION 'representation symbol is not attached to the revision';
-                    END IF;
-                    IF NEW.footprint_asset_id IS NOT NULL AND NOT EXISTS (
-                        SELECT 1 FROM revision_assets link JOIN assets asset ON asset.id = link.asset_id
-                        WHERE link.revision_id = NEW.revision_id AND link.asset_id = NEW.footprint_asset_id
-                          AND link.asset_type = 'footprint' AND asset.asset_type = 'footprint'
-                    ) THEN
-                        RAISE EXCEPTION 'representation footprint is not attached to the revision';
-                    END IF;
-                    RETURN NEW;
-                END;
-                $$
-                """
-            )
-            conn.execute(
-                """
-                CREATE OR REPLACE FUNCTION prism_validate_revision_release()
-                RETURNS trigger
-                LANGUAGE plpgsql
-                AS $$
-                DECLARE
-                    component_identity_kind TEXT;
-                    complete_defaults INTEGER;
-                BEGIN
-                    IF NEW.release_status NOT IN ('done', 'released') THEN
-                        RETURN NEW;
-                    END IF;
-                    SELECT identity_kind INTO component_identity_kind
-                    FROM components WHERE id = NEW.component_id;
-                    IF component_identity_kind = 'provisional_ipn' THEN
-                        RAISE EXCEPTION 'provisional components cannot reach done or released';
-                    END IF;
-                    IF NEW.release_status = 'released' THEN
-                        SELECT COUNT(*) INTO complete_defaults
-                        FROM revision_representations
-                        WHERE revision_id = NEW.id AND is_default = 1
-                          AND symbol_asset_id IS NOT NULL AND footprint_asset_id IS NOT NULL;
-                    END IF;
-                    IF NEW.release_status = 'released' AND complete_defaults <> 1 THEN
-                        RAISE EXCEPTION 'released revisions require exactly one complete default representation';
-                    END IF;
-                    RETURN NEW;
-                END;
-                $$
-                """
-            )
-            asset_update_trigger = conn.execute(
-                """
-                SELECT 1 AS present
-                FROM pg_trigger
-                WHERE tgname = 'trg_assets_identity_update'
-                  AND tgrelid = to_regclass('catalog.assets')
-                  AND NOT tgisinternal
-                """
-            ).fetchone()
-            if not asset_update_trigger:
-                conn.execute(
-                    "CREATE TRIGGER trg_assets_identity_update BEFORE UPDATE ON assets "
-                    "FOR EACH ROW EXECUTE FUNCTION prism_guard_asset_identity_update()"
-                )
-            revision_update_trigger = conn.execute(
-                """
-                SELECT 1 AS present
-                FROM pg_trigger
-                WHERE tgname = 'trg_component_revisions_finalized_update'
-                  AND tgrelid = to_regclass('catalog.component_revisions')
-                  AND NOT tgisinternal
-                """
-            ).fetchone()
-            if not revision_update_trigger:
-                conn.execute(
-                    "CREATE TRIGGER trg_component_revisions_finalized_update "
-                    "BEFORE UPDATE ON component_revisions "
-                    "FOR EACH ROW EXECUTE FUNCTION prism_guard_finalized_revision_update()"
-                )
-            revision_preview_trigger = conn.execute(
-                """
-                SELECT 1 AS present
-                FROM pg_trigger
-                WHERE tgname = 'trg_revision_previews_finalized'
-                  AND tgrelid = to_regclass('catalog.revision_previews')
-                  AND NOT tgisinternal
-                """
-            ).fetchone()
-            if not revision_preview_trigger:
-                conn.execute(
-                    "CREATE TRIGGER trg_revision_previews_finalized "
-                    "BEFORE INSERT OR UPDATE OR DELETE ON revision_previews "
-                    "FOR EACH ROW EXECUTE FUNCTION prism_guard_revision_preview_mutation()"
-                )
-            revision_asset_trigger = conn.execute(
-                """
-                SELECT 1 AS present
-                FROM pg_trigger
-                WHERE tgname = 'trg_revision_assets_finalized'
-                  AND tgrelid = to_regclass('catalog.revision_assets')
-                  AND NOT tgisinternal
-                """
-            ).fetchone()
-            if not revision_asset_trigger:
-                conn.execute(
-                    "CREATE TRIGGER trg_revision_assets_finalized "
-                    "BEFORE INSERT OR UPDATE OR DELETE ON revision_assets "
-                    "FOR EACH ROW EXECUTE FUNCTION prism_guard_revision_preview_mutation()"
-                )
-            for trigger_name, table, operations, function_name in (
-                (
-                    "trg_revision_assets_type",
-                    "revision_assets",
-                    "INSERT OR UPDATE",
-                    "prism_validate_revision_asset_type",
-                ),
-                (
-                    "trg_revision_representations_finalized",
-                    "revision_representations",
-                    "INSERT OR UPDATE OR DELETE",
-                    "prism_guard_revision_preview_mutation",
-                ),
-                (
-                    "trg_revision_representations_assets",
-                    "revision_representations",
-                    "INSERT OR UPDATE",
-                    "prism_validate_representation_assets",
-                ),
-                (
-                    "trg_component_revisions_release",
-                    "component_revisions",
-                    "INSERT OR UPDATE OF release_status",
-                    "prism_validate_revision_release",
-                ),
-            ):
-                conn.execute(f"DROP TRIGGER IF EXISTS {trigger_name} ON {table}")
-                conn.execute(
-                    f"CREATE TRIGGER {trigger_name} BEFORE {operations} ON {table} "
-                    f"FOR EACH ROW EXECUTE FUNCTION {function_name}()"
-                )
-            conn.execute(
-                """
-                INSERT INTO catalog_meta (key, value) VALUES (%s, %s)
-                ON CONFLICT (key) DO UPDATE SET value = EXCLUDED.value
-                """,
-                ("postgres_integrity_guards_version", POSTGRES_INTEGRITY_GUARDS_VERSION),
-            )
-            conn.commit()
+        ensure_postgres_integrity_guards(self._postgres_runtime)
 
     def _clone_revision(self, conn: Any, component_id: str, **kwargs: Any) -> dict[str, Any]:
         # Serialize version allocation and head updates per component. The unique
